+++ ext/puma_http11/puma_http11.c
@@ -434,7 +434,6 @@ VALUE HttpParser_body(VALUE self) {
   return http->body;
 }
 
-void Init_io_buffer(VALUE puma);
 void Init_mini_ssl(VALUE mod);
 
 void Init_puma_http11()
@@ -464,6 +463,5 @@ void Init_puma_http11()
   rb_define_method(cHttpParser, "body", HttpParser_body, 0);
   init_common_fields();
 
-  Init_io_buffer(mPuma);
   Init_mini_ssl(mPuma);
 }