+++ History.md
@@ -3,18 +3,38 @@
 * Features
   * Add pumactl `thread-backtraces` command to print thread backtraces (#2053)
   * Configuration: `environment` is read from `RAILS_ENV`, if `RACK_ENV` can't be found (#2022)
+  * Do not set user_config to quiet by default to allow for file config (#2074)
   * `Puma.stats` now returns a Hash instead of a JSON string (#2086)
   * `GC.compact` is called before fork if available (#2093)
+  * Add `requests_count` to workers stats. (#2106)
+  * Changed #connected_port to #connected_ports (#2076)
+  * `--control` has been removed. Use `--control-url` (#1487)
+  * `worker_directory` has been removed. Use `directory`
 
 * Bugfixes
-  * Your bugfix goes here (#Github Number)
   * Windows update extconf.rb for use with ssp and varied Ruby/MSYS2 combinations (#2069)
   * Ensure control server Unix socket is closed on shutdown (#2112)
+  * Preserve `BUNDLE_GEMFILE` env var when using `prune_bundler` (#1893)
+  * Send 408 request timeout even when queue requests is disabled (#2119)
+  * Rescue IO::WaitReadable instead of EAGAIN for blocking read (#2121)
 
 * Refactor
   * Remove unused loader argument from Plugin initializer (#2095)
   * Simplify `Configuration.random_token` and remove insecure fallback (#2102)
   * Simplify `Runner#start_control` URL parsing (#2111)
+  * Removed the IOBuffer extension and replaced with Ruby (#1980)
+
+
+## 4.3.3 and 3.12.4 / 2020-02-28
+  * Bugfixes
+    * Fix: Fixes a problem where we weren't splitting headers correctly on newlines (#2132)
+  * Security
+    * Fix: Prevent HTTP Response splitting via CR in early hints.
+
+## 4.3.2 and 3.12.3 / 2020-02-27
+
+* Security
+  * Fix: Prevent HTTP Response splitting via CR/LF in header values. CVE-2020-5247.
 
 ## 4.3.1 and 3.12.2 / 2019-12-05